PM-1377 Refactoring
