+++ src/main/java/it/pagopa/pm/gateway/service/CcResumeStep2Service.java
@@ -4,22 +4,17 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import it.pagopa.pm.gateway.client.ecommerce.EcommerceClient;
 import it.pagopa.pm.gateway.client.vpos.HttpClient;
 import it.pagopa.pm.gateway.client.vpos.HttpClientResponse;
-import it.pagopa.pm.gateway.dto.config.ClientConfig;
 import it.pagopa.pm.gateway.dto.creditcard.StepZeroRequest;
-import it.pagopa.pm.gateway.dto.enums.RefundOutcome;
 import it.pagopa.pm.gateway.dto.enums.ThreeDS2ResponseTypeEnum;
-import it.pagopa.pm.gateway.dto.transaction.AuthResultEnum;
-import it.pagopa.pm.gateway.dto.transaction.TransactionInfo;
-import it.pagopa.pm.gateway.dto.transaction.UpdateAuthRequest;
 import it.pagopa.pm.gateway.dto.vpos.AuthResponse;
 import it.pagopa.pm.gateway.dto.vpos.ThreeDS2Authorization;
 import it.pagopa.pm.gateway.dto.vpos.ThreeDS2Response;
-import it.pagopa.pm.gateway.dto.vpos.VposOrderStatusResponse;
 import it.pagopa.pm.gateway.entity.PaymentRequestEntity;
 import it.pagopa.pm.gateway.repository.PaymentRequestRepository;
 import it.pagopa.pm.gateway.utils.ClientsConfig;
 import it.pagopa.pm.gateway.utils.VPosRequestUtils;
 import it.pagopa.pm.gateway.utils.VPosResponseUtils;
+import it.pagopa.pm.gateway.utils.VposPatchUtils;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -35,17 +30,14 @@ import java.util.Arrays;
 import java.util.Map;
 import java.util.Objects;
 
-import static it.pagopa.pm.gateway.constant.Messages.*;
+import static it.pagopa.pm.gateway.constant.Messages.GENERIC_ERROR_MSG;
 import static it.pagopa.pm.gateway.constant.VposConstant.RESULT_CODE_AUTHORIZED;
 import static it.pagopa.pm.gateway.dto.enums.PaymentRequestStatusEnum.*;
-import static it.pagopa.pm.gateway.dto.enums.RefundOutcome.KO;
-import static it.pagopa.pm.gateway.dto.enums.RefundOutcome.OK;
 
 @Service
 @Slf4j
 @NoArgsConstructor
 public class CcResumeStep2Service {
-    private ClientsConfig clientsConfig;
 
     @Value("${vpos.requestUrl}")
     private String vposUrl;
@@ -69,9 +61,7 @@ public class CcResumeStep2Service {
     private ObjectMapper objectMapper;
 
     @Autowired
-    public CcResumeStep2Service(ClientsConfig clientsConfig) {
-        this.clientsConfig = clientsConfig;
-    }
+    private VposPatchUtils vposPatchUtils;
 
     public void startResumeStep2(String requestId) {
         PaymentRequestEntity entity = paymentRequestRepository.findByGuid(requestId);
@@ -116,7 +106,7 @@ public class CcResumeStep2Service {
                 executeAccount(entity, request);
             }
 
-            executePatchTransaction(entity, request);
+            vposPatchUtils.executePatchTransaction(entity, request);
         } catch (Exception e) {
             log.error("{}{}", GENERIC_ERROR_MSG, entity.getIdTransaction(), e);
         }
@@ -183,76 +173,5 @@ public class CcResumeStep2Service {
         entity.setErrorCode(errorCode);
         paymentRequestRepository.save(entity);
         log.info("END - XPay Request Payment Account for requestId {}", entity.getGuid());
-    }
-
-    private void executePatchTransaction(PaymentRequestEntity entity, StepZeroRequest pgsRequest) throws IOException {
-        String requestId = entity.getGuid();
-        log.info("START - PATCH updateTransaction for requestId: {}", requestId);
-        AuthResultEnum authResult = entity.getStatus().equals(AUTHORIZED.name()) ? AuthResultEnum.OK : AuthResultEnum.KO;
-
-        String authCode;
-        if (AUTHORIZED.name().equals(entity.getStatus())) {
-            authCode = entity.getAuthorizationCode();
-        } else {
-            authCode = entity.getErrorCode();
-        }
-
-        UpdateAuthRequest patchRequest = new UpdateAuthRequest(authResult, authCode);
-        try {
-            ClientConfig clientConfig = clientsConfig.getByKey(entity.getClientId());
-            TransactionInfo patchResponse = ecommerceClient.callPatchTransaction(patchRequest, entity.getIdTransaction(), clientConfig);
-            log.info("Response from PATCH updateTransaction for requestId {} is {}", requestId, patchResponse.toString());
-        } catch (Exception e) {
-            log.error("{}{}", PATCH_CLOSE_PAYMENT_ERROR, requestId, e);
-            log.info("Refunding payment with requestId: {}", requestId);
-            if (executeOrderStatus(entity, pgsRequest).equals(OK)) {
-                executeRevert(entity, pgsRequest);
-            }
-        }
-    }
-
-    private RefundOutcome executeOrderStatus(PaymentRequestEntity entity, StepZeroRequest stepZeroRequest) throws IOException {
-        log.info("Calling VPOS - OrderStatus - for requestId: {}", entity.getGuid());
-        Map<String, String> params = vPosRequestUtils.buildOrderStatusParams(stepZeroRequest);
-        HttpClientResponse clientResponse = callVPos(params);
-        VposOrderStatusResponse response = vPosResponseUtils.buildOrderStatusResponse(clientResponse.getEntity());
-        return computeOrderStatusResultCode(response, entity);
-    }
-
-    private RefundOutcome computeOrderStatusResultCode(VposOrderStatusResponse response, PaymentRequestEntity entity) {
-        String resultCode = response.getResultCode();
-        if (resultCode.equals(RESULT_CODE_AUTHORIZED)) {
-            return OK;
-        } else {
-            entity.setErrorMessage("Error during orderStatus");
-            paymentRequestRepository.save(entity);
-            return KO;
-        }
-    }
-
-    private void executeRevert(PaymentRequestEntity entity, StepZeroRequest pgsRequest) {
-        try {
-            log.info("Calling VPOS - Revert - for requestId: " + entity.getGuid());
-            Map<String, String> params = vPosRequestUtils.buildRevertRequestParams(pgsRequest, entity.getCorrelationId());
-            HttpClientResponse clientResponse = callVPos(params);
-            AuthResponse response = vPosResponseUtils.buildAuthResponse(clientResponse.getEntity());
-            vPosResponseUtils.validateResponseMac(response.getTimestamp(), response.getResultCode(), response.getResultMac(), pgsRequest);
-            checkRevertResultCode(response, entity);
-        } catch (Exception e) {
-            log.error("{}{} cause: {} - {}", GENERIC_REFUND_ERROR_MSG, entity.getIdTransaction(), e.getCause(),  e.getMessage(), e);
-        }
-    }
-
-    private void checkRevertResultCode(AuthResponse response, PaymentRequestEntity entity) {
-        String resultCode = response.getResultCode();
-        if (resultCode.equals(RESULT_CODE_AUTHORIZED)) {
-            entity.setStatus(CANCELLED.name());
-            entity.setIsRefunded(true);
-            paymentRequestRepository.save(entity);
-        } else {
-            entity.setErrorMessage("Error during Revert");
-            entity.setIsRefunded(false);
-        }
-        log.info("END - VPos Request Payment Revert for requestId " + entity.getGuid() + " - resultCode: " + resultCode);
     }
 }
