+++ src/test/java/it/pagopa/pm/gateway/service/CcResumeStep1ServiceTest.java
@@ -9,13 +9,10 @@ import it.pagopa.pm.gateway.dto.config.VposClientConfig;
 import it.pagopa.pm.gateway.dto.creditcard.CreditCardResumeRequest;
 import it.pagopa.pm.gateway.dto.creditcard.StepZeroRequest;
 import it.pagopa.pm.gateway.dto.enums.ThreeDS2ResponseTypeEnum;
-import it.pagopa.pm.gateway.dto.transaction.TransactionInfo;
 import it.pagopa.pm.gateway.dto.vpos.AuthResponse;
 import it.pagopa.pm.gateway.dto.vpos.ThreeDS2Response;
-import it.pagopa.pm.gateway.dto.vpos.VposOrderStatusResponse;
 import it.pagopa.pm.gateway.entity.PaymentRequestEntity;
 import it.pagopa.pm.gateway.repository.PaymentRequestRepository;
-import it.pagopa.pm.gateway.utils.ClientsConfig;
 import it.pagopa.pm.gateway.utils.VPosRequestUtils;
 import it.pagopa.pm.gateway.utils.VPosResponseUtils;
 import org.junit.Before;
@@ -39,19 +36,17 @@ import static org.mockito.Mockito.when;
 @RunWith(MockitoJUnitRunner.class)
 @SpringBootTest(classes = CcResumeStep1Service.class)
 public class CcResumeStep1ServiceTest {
+    public static final String RESULT_CODE_00 = "00";
     private final ClientConfig clientConfig = new ClientConfig();
-
-    @Mock
-    private ClientsConfig clientsConfig;
 
     @Spy
     @InjectMocks
-    private CcResumeStep1Service service = new CcResumeStep1Service(clientsConfig);
+    private CcResumeStep1Service service;
 
     @Before
     public void setUpProperties() {
         ReflectionTestUtils.setField(service, "vposUrl", "http://localhost:8080");
-        VposClientConfig vposClientConfig= new VposClientConfig();
+        VposClientConfig vposClientConfig = new VposClientConfig();
         vposClientConfig.setClientReturnUrl("url");
         clientConfig.setVpos(vposClientConfig);
     }
@@ -102,11 +97,10 @@ public class CcResumeStep1ServiceTest {
         entity.setIdTransaction("1234566");
 
         ThreeDS2Response response = ValidBeans.createThreeDS2ResponseStep0Authorization();
-        AuthResponse authResponse = ValidBeans.createVPosAuthResponse("00");
+        AuthResponse authResponse = ValidBeans.createVPosAuthResponse(RESULT_CODE_00);
         Map<String, String> params = new HashMap<>();
         params.put("1", "prova");
 
-        when(clientsConfig.getByKey(any())).thenReturn(clientConfig);
         when(objectMapper.readValue(entity.getJsonRequest(), StepZeroRequest.class)).thenReturn(stepZeroRequest);
         when(vPosRequestUtils.buildStepOneRequestParams(any(), any(), any())).thenReturn(params);
         when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
@@ -114,7 +108,6 @@ public class CcResumeStep1ServiceTest {
         when((vPosRequestUtils.buildAccountingRequestParams(any(), any()))).thenReturn(params);
         when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
         when(vPosResponseUtils.buildAuthResponse(any())).thenReturn(authResponse);
-        when(ecommerceClient.callPatchTransaction(any(), any(), any())).thenReturn(new TransactionInfo());
 
         when(paymentRequestRepository.findByGuid(any())).thenReturn(entity);
         service.startResumeStep1(creditCardResumeRequest, UUID_SAMPLE);
@@ -255,7 +248,6 @@ public class CcResumeStep1ServiceTest {
         Map<String, String> params = new HashMap<>();
         params.put("1", "prova");
 
-        when(clientsConfig.getByKey(any())).thenReturn(clientConfig);
         when(objectMapper.readValue(entity.getJsonRequest(), StepZeroRequest.class)).thenReturn(stepZeroRequest);
         when(vPosRequestUtils.buildStepOneRequestParams(any(), any(), any())).thenReturn(params);
         when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
@@ -263,7 +255,6 @@ public class CcResumeStep1ServiceTest {
         when((vPosRequestUtils.buildAccountingRequestParams(any(), any()))).thenReturn(params);
         when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
         when(vPosResponseUtils.buildAuthResponse(any())).thenThrow(RuntimeException.class);
-        when(ecommerceClient.callPatchTransaction(any(), any(), any())).thenReturn(new TransactionInfo());
 
         when(paymentRequestRepository.findByGuid(any())).thenReturn(entity);
         service.startResumeStep1(creditCardResumeRequest, UUID_SAMPLE);
@@ -287,7 +278,6 @@ public class CcResumeStep1ServiceTest {
         Map<String, String> params = new HashMap<>();
         params.put("1", "prova");
 
-        when(clientsConfig.getByKey(any())).thenReturn(clientConfig);
         when(objectMapper.readValue(entity.getJsonRequest(), StepZeroRequest.class)).thenReturn(stepZeroRequest);
         when(vPosRequestUtils.buildStepOneRequestParams(any(), any(), any())).thenReturn(params);
         when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
@@ -295,77 +285,6 @@ public class CcResumeStep1ServiceTest {
         when((vPosRequestUtils.buildAccountingRequestParams(any(), any()))).thenReturn(params);
         when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
         when(vPosResponseUtils.buildAuthResponse(any())).thenReturn(authResponse);
-        when(ecommerceClient.callPatchTransaction(any(), any(), any())).thenReturn(new TransactionInfo());
-
-        when(paymentRequestRepository.findByGuid(any())).thenReturn(entity);
-        service.startResumeStep1(creditCardResumeRequest, UUID_SAMPLE);
-        verify(service).startResumeStep1(creditCardResumeRequest, UUID_SAMPLE);
-    }
-
-    @Test
-    public void startResume_Exception_in_executePatchTransaction() throws IOException {
-        StepZeroRequest stepZeroRequest = ValidBeans.createStep0Request(false);
-        CreditCardResumeRequest creditCardResumeRequest = ValidBeans.createCreditCardResumeRequest(true);
-
-        PaymentRequestEntity entity = new PaymentRequestEntity();
-        entity.setResponseType(ThreeDS2ResponseTypeEnum.METHOD.name());
-        String requestJson = objectMapper.writeValueAsString(stepZeroRequest);
-        entity.setJsonRequest(requestJson);
-        entity.setCorrelationId("CorrelationId");
-        entity.setIdTransaction("1234566");
-
-        ThreeDS2Response response = ValidBeans.createThreeDS2ResponseStep0Authorization();
-        AuthResponse authResponse = ValidBeans.createVPosAuthResponse("00");
-        Map<String, String> params = new HashMap<>();
-        params.put("1", "prova");
-        VposOrderStatusResponse vposOrderStatusResponse = new VposOrderStatusResponse();
-        vposOrderStatusResponse.setResultCode("00");
-
-        when(clientsConfig.getByKey(any())).thenReturn(clientConfig);
-        when(objectMapper.readValue(entity.getJsonRequest(), StepZeroRequest.class)).thenReturn(stepZeroRequest);
-        when(vPosRequestUtils.buildStepOneRequestParams(any(), any(), any())).thenReturn(params);
-        when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
-        when(vPosResponseUtils.build3ds2Response(any())).thenReturn(response);
-        when((vPosRequestUtils.buildAccountingRequestParams(any(), any()))).thenReturn(params);
-        when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
-        when(vPosResponseUtils.buildAuthResponse(any())).thenReturn(authResponse);
-        when(ecommerceClient.callPatchTransaction(any(), any(), any())).thenThrow(RuntimeException.class);
-        when(vPosResponseUtils.buildOrderStatusResponse(any())).thenReturn(vposOrderStatusResponse);
-
-        when(paymentRequestRepository.findByGuid(any())).thenReturn(entity);
-        service.startResumeStep1(creditCardResumeRequest, UUID_SAMPLE);
-        verify(service).startResumeStep1(creditCardResumeRequest, UUID_SAMPLE);
-    }
-
-    @Test
-    public void startResume_Exception_in_executePatchTransaction_And_Error_In_OrderStatus() throws IOException {
-        StepZeroRequest stepZeroRequest = ValidBeans.createStep0Request(false);
-        CreditCardResumeRequest creditCardResumeRequest = ValidBeans.createCreditCardResumeRequest(true);
-
-        PaymentRequestEntity entity = new PaymentRequestEntity();
-        entity.setResponseType(ThreeDS2ResponseTypeEnum.METHOD.name());
-        String requestJson = objectMapper.writeValueAsString(stepZeroRequest);
-        entity.setJsonRequest(requestJson);
-        entity.setCorrelationId("CorrelationId");
-        entity.setIdTransaction("1234566");
-
-        ThreeDS2Response response = ValidBeans.createThreeDS2ResponseStep0Authorization();
-        AuthResponse authResponse = ValidBeans.createVPosAuthResponse("00");
-        Map<String, String> params = new HashMap<>();
-        params.put("1", "prova");
-        VposOrderStatusResponse vposOrderStatusResponse = new VposOrderStatusResponse();
-        vposOrderStatusResponse.setResultCode("02");
-
-        when(clientsConfig.getByKey(any())).thenReturn(clientConfig);
-        when(objectMapper.readValue(entity.getJsonRequest(), StepZeroRequest.class)).thenReturn(stepZeroRequest);
-        when(vPosRequestUtils.buildStepOneRequestParams(any(), any(), any())).thenReturn(params);
-        when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
-        when(vPosResponseUtils.build3ds2Response(any())).thenReturn(response);
-        when((vPosRequestUtils.buildAccountingRequestParams(any(), any()))).thenReturn(params);
-        when(httpClient.post(any(), any(), any())).thenReturn(ValidBeans.createHttpClientResponseVPos());
-        when(vPosResponseUtils.buildAuthResponse(any())).thenReturn(authResponse);
-        when(ecommerceClient.callPatchTransaction(any(), any(), any())).thenThrow(RuntimeException.class);
-        when(vPosResponseUtils.buildOrderStatusResponse(any())).thenReturn(vposOrderStatusResponse);
 
         when(paymentRequestRepository.findByGuid(any())).thenReturn(entity);
         service.startResumeStep1(creditCardResumeRequest, UUID_SAMPLE);
